+++ src/commands/funnies/cowsay.ts
@@ -1,18 +1,27 @@
+import Filter from 'bad-words';
 import { say } from 'cowsay';
 import { CommandDefinition } from '../../lib/command';
-import { CommandCategory } from '../../constants';
+import { Channels, CommandCategory } from '../../constants';
 
 export const cowsay: CommandDefinition = {
     name: ['cowsay', 'cs'],
     description: 'Emulates the famous UNIX program `cowsay`.',
     category: CommandCategory.FUNNIES,
     executor: (msg) => {
-        const text = msg.content.replace(/\.(cowsay|cs)\s*/, '');
+        const filter = new Filter();
+        const text = msg.content.replace(/\.(cowsay|cs)\s/, '').replace(/`/g, '');
+        if (msg.channel.id !== Channels.BOT_COMMANDS) {
+            return msg.reply(`This command can only be used in <#${Channels.BOT_COMMANDS}>`);
+        }
+
+        if (filter.isProfane(text)) {
+            return msg.reply('Please do not use profane language with this command.');
+        }
 
         if (text) {
             return msg.channel.send(`\`\`\`\n${say(({ text }))}\n\`\`\``);
         }
 
-        return msg.reply('please provide some text.');
+        return msg.reply('Please provide some text.');
     },
 };